Revert "phraseRendering.tsの処理をsinging.tsに移動"
This reverts commit 8247210052b7390f08144d54aba78597eba6b3d4.
